refactor: :art: fixed imports addresses, fixed login error handling
